# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EmoFi Protocol API — Reality-Integrated Emotional Finance
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create or register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: Wallet address already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}/portfolio:
    get:
      operationId: getUserPortfolio
      tags: [users]
      summary: Get full portfolio — balances, vaults, staking
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: User portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPortfolio"
  /users/{userId}/activity:
    get:
      operationId: getUserActivity
      tags: [users]
      summary: Recent activity feed for user
      parameters:
        - $ref: "#/components/parameters/userId"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Activity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # ─── TOKENS ───────────────────────────────────────────────────────────────
...
  /vaults:
    get:
      operationId: listVaults
      tags: [vaults]
      summary: List vaults, optionally filtered by user or token type
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: tokenType
          in: query
          schema:
            $ref: "#/components/schemas/TokenType"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Vault list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultList"

    post:
      operationId: createVault
      tags: [vaults]
      summary: Create a new RI-Vault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVaultBody"
      responses:
        "201":
          description: Vault created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"
  /vaults/{vaultId}:
    get:
      operationId: getVault
      tags: [vaults]
      summary: Get vault details with recent transactions and oracle feeds
      parameters:
        - $ref: "#/components/parameters/vaultId"
      responses:
        "200":
          description: Vault details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vaults/{vaultId}/mint:
    post:
      operationId: mintToVault
      tags: [vaults]
      summary: Mint attribute tokens into a vault
      parameters:
        - $ref: "#/components/parameters/vaultId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultMintBody"
      responses:
        "200":
          description: Mint result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultTransaction"
  /vaults/{vaultId}/burn:
    post:
      operationId: burnFromVault
      tags: [vaults]
      summary: Burn tokens from a vault
      parameters:
        - $ref: "#/components/parameters/vaultId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultBurnBody"
      responses:
        "200":
          description: Burn result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultTransaction"
  /vaults/{vaultId}/transactions:
    get:
      operationId: getVaultTransactions
      tags: [vaults]
      summary: Transaction history for a vault
      parameters:
        - $ref: "#/components/parameters/vaultId"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VaultTransaction"

  # ─── MARKETPLACE ──────────────────────────────────────────────────────────
...
  /staking/positions:
    get:
      operationId: listStakingPositions
      tags: [staking]
      summary: List staking positions for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staking positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StakingPosition"

    post:
      operationId: stakeTokens
      tags: [staking]
      summary: Stake tokens to earn rewards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StakeBody"
      responses:
        "201":
          description: Staking position created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StakingPosition"
  /staking/positions/{positionId}/unstake:
    post:
      operationId: unstakeTokens
      tags: [staking]
      summary: Unstake and collect rewards
      parameters:
        - $ref: "#/components/parameters/positionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnstakeBody"
      responses:
        "200":
          description: Unstake result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StakingPosition"
  /staking/positions/{positionId}/claim:
    post:
      operationId: claimRewards
      tags: [staking]
      summary: Claim pending staking rewards
      parameters:
        - $ref: "#/components/parameters/positionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimBody"
      responses:
        "200":
          description: Claim result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardClaim"
...
components:
  parameters:
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
    vaultId:
      name: vaultId
      in: path
      required: true
      schema:
        type: integer
...
    positionId:
      name: positionId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    TokenType:
      type: string
      enum:
        - happiness
        - sadness
        - beautiful
        - good_thought
        - bad_thought
        - intelligence
        - talent
        - spirituality
        - situational
    TransactionType:
      type: string
      enum: [mint, burn, stake, unstake, transfer, buy, sell]
...
    OracleFeedType:
      type: string
      enum:
        - mood
        - achievement
        - talent_score
        - intelligence_score
        - spiritual_index
        - situational_event
...
    User:
      type: object
      properties:
        id:
          type: integer
        walletAddress:
          type: string
        username:
          type: string
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        emoBalance:
          type: number
        reputationScore:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, walletAddress, username, emoBalance, reputationScore, createdAt]
    CreateUserBody:
      type: object
      properties:
        walletAddress:
          type: string
        username:
          type: string
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
      required: [walletAddress, username]
    UserPortfolio:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        tokenBalances:
          type: array
          items:
            $ref: "#/components/schemas/UserTokenBalance"
        vaults:
          type: array
          items:
            $ref: "#/components/schemas/Vault"
        stakingPositions:
          type: array
          items:
            $ref: "#/components/schemas/StakingPosition"
        totalPortfolioValue:
          type: number
        pendingRewards:
          type: number
      required: [user, tokenBalances, vaults, stakingPositions, totalPortfolioValue, pendingRewards]
    UserTokenBalance:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        tokenType:
          $ref: "#/components/schemas/TokenType"
        balance:
          type: number
        lockedBalance:
          type: number
      required: [id, userId, tokenType, balance, lockedBalance]
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        txType:
          $ref: "#/components/schemas/TransactionType"
        tokenType:
          $ref: "#/components/schemas/TokenType"
          nullable: true
        amount:
          type: number
          nullable: true
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, txType, description, createdAt]

    # ── Tokens ────────────────────────────────────────────────────────────
...
    Vault:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        tokenType:
          $ref: "#/components/schemas/TokenType"
        description:
          type: string
          nullable: true
        balance:
          type: number
        stakedBalance:
          type: number
        oracleScore:
          type: number
          nullable: true
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, name, tokenType, balance, stakedBalance, isPublic, createdAt, updatedAt]
    VaultDetail:
      allOf:
        - $ref: "#/components/schemas/Vault"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            recentTransactions:
              type: array
              items:
                $ref: "#/components/schemas/VaultTransaction"
            oracleFeeds:
              type: array
              items:
                $ref: "#/components/schemas/OracleFeed"
          required: [user, recentTransactions, oracleFeeds]
    VaultList:
      type: object
      properties:
        vaults:
          type: array
          items:
            $ref: "#/components/schemas/Vault"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
      required: [vaults, total, offset, limit]
    CreateVaultBody:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        tokenType:
          $ref: "#/components/schemas/TokenType"
        description:
          type: string
          nullable: true
        isPublic:
          type: boolean
          default: true
      required: [userId, name, tokenType]
    VaultMintBody:
      type: object
      properties:
        userId:
          type: integer
        amount:
          type: number
        note:
          type: string
          nullable: true
      required: [userId, amount]
    VaultBurnBody:
      type: object
      properties:
        userId:
          type: integer
        amount:
          type: number
        reason:
          type: string
          nullable: true
      required: [userId, amount]
    VaultTransaction:
      type: object
      properties:
        id:
          type: integer
        vaultId:
          type: integer
        userId:
          type: integer
        txType:
          $ref: "#/components/schemas/TransactionType"
        tokenType:
          $ref: "#/components/schemas/TokenType"
        amount:
          type: number
        balanceAfter:
          type: number
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, vaultId, userId, txType, tokenType, amount, balanceAfter, createdAt]

    # ── Marketplace ───────────────────────────────────────────────────────
...
    StakingPosition:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        stakedTokenType:
          $ref: "#/components/schemas/TokenType"
        rewardTokenType:
          $ref: "#/components/schemas/TokenType"
        amountStaked:
          type: number
        rewardRate:
          type: number
        pendingRewards:
          type: number
        totalEarned:
          type: number
        isActive:
          type: boolean
        startedAt:
          type: string
          format: date-time
        lastClaimedAt:
          type: string
          format: date-time
          nullable: true
        unstakedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, stakedTokenType, rewardTokenType, amountStaked, rewardRate, pendingRewards, totalEarned, isActive, startedAt]
    StakeBody:
      type: object
      properties:
        userId:
          type: integer
        stakedTokenType:
          $ref: "#/components/schemas/TokenType"
        rewardTokenType:
          $ref: "#/components/schemas/TokenType"
        amount:
          type: number
      required: [userId, stakedTokenType, rewardTokenType, amount]
    UnstakeBody:
      type: object
      properties:
        userId:
          type: integer
      required: [userId]
    ClaimBody:
      type: object
      properties:
        userId:
          type: integer
      required: [userId]
    RewardClaim:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        positionId:
          type: integer
          nullable: true
        amount:
          type: number
        tokenType:
          $ref: "#/components/schemas/TokenType"
        source:
          type: string
        claimedAt:
          type: string
          format: date-time
      required: [id, userId, amount, tokenType, source, claimedAt]
...
    OracleFeed:
      type: object
      properties:
        id:
          type: integer
        feedType:
          $ref: "#/components/schemas/OracleFeedType"
        userId:
          type: integer
          nullable: true
        value:
          type: number
        source:
          type: string
        verified:
          type: boolean
        metadata:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, feedType, value, source, verified, createdAt]
...